이동 동기화 클라 권위 섞었다. 클라 측 충돌 처리 중.

+++ Server/Client/DefaultWindow/Types.h
@@ -121,11 +121,11 @@ struct VectorInt
 		return ret;
 	}
 
-	VectorInt operator*(int32 value)
+	VectorInt operator*(float value)
 	{
 		VectorInt ret;
-		ret.x = x * value;
-		ret.y = y * value;
+		ret.x = (int32)(x * value);
+		ret.y = (int32)(y * value);
 		return ret;
 	}
